+++ app-ios/Packages/library-common/Sources/Setting/Repositories/Implementations/DefaultSettingRepository.swift
@@ -36,7 +36,7 @@ struct DefaultSettingRepository: SettingRepository {
 
         return SettingSnapshot(
             accountInfo: SettingAccountInfo(
-                nickname: userInfo?.nikeName,
+                nickname: userInfo?.nickname,
                 username: userInfo?.username,
                 phone: userInfo?.phone
             ),